Add a `requirements.txt` file for one-step install
This allows users who want to install all Python package dependencies in
a single step to do so simply with `pip install -r requirements.txt`.
Of course, anyone without `pip` or who wants to install each dependency
manually is still able to do so by simply ignoring the file.

Note that this commit assumes that the `astrolib.coords` package that is
now on PyPI is the same package that `astrolib/convert.py` intends to
test, even though that code previously seems to have been able to simply
import the package name `coords` without needing to prefix it.  This
might indicate that I have gotten two similarly-named packages confused
and need to adjust that dependency and restore the original form of the
import statement.