Add palpy to coordinate testing
palpy is a python wrapper for the C PAL library which is itself
a wrapper around SOFA/ERFA. It uses the SLALIB API but does not
use any Fortran. Where SOFA/ERFA lacks functionality the Fortran
SLALIB source has been ported to C.

The conversion script is a straight port of the pyslalib test
script.